Add biomedical and natural image masking (#47)
* Add biomedical and natural image masking

In this commit, masking routins are added to give users the ability
to join two tables and perform image-specific masking in both the
image and biomedimage action sets. Test cases are added for
decoded and encoded situations.

* Add biomedical and natural image masking

In this commit, masking routins are added to give users the ability
to join two tables and perform image-specific masking in both the
image and biomedimage action sets. Test cases are added for
decoded and encoded situations.